refactor(docs): updated styling and decoupling
