+++ docs/src/pages/templates.module.css
@@ -70,12 +70,16 @@ body {
 	width: 100%;
 	height: 100%;
 	border-color: rgba(255, 255, 255, 0.12);
-	border-width: 1px;
+	border-width: 2px;
 	background-color: transparent;
-	outline-color: var(--ifm-color-primary);
-	color: white;
+	outline: 2px solid transparent;
+	transition: outline-color 0.3s ease;
 	padding-left: 40px;
 	padding-right: 12px;
+}
+
+.searchInput:focus {
+	outline-color: var(--ifm-color-primary);
 }
 
 .filterOptions {
@@ -93,12 +97,21 @@ body {
 	flex-direction: row;
 	border-radius: 0px;
 	border-width: 0px;
+	transition: background-color 0.3s ease-in-out;
 	align-items: center;
 	cursor: pointer;
 }
 
+.filterOptionUnselected:hover {
+	background-color: var(--hover-background-color);
+}
+
 .filterOptionSelected {
-	background-color: RGB(0, 121, 107, .24);
+	background-color: RGB(0, 121, 107, 0.24);
+}
+
+.filterOptionSelected:hover {
+	background-color: RGB(0, 121, 107, 0.5);
 }
 
 .filterOptionLabel {
@@ -109,7 +122,7 @@ body {
 }
 
 .filterOptionLabelSelected {
-	color: #00BFA5;
+	color: #00bfa5;
 	font-weight: 600;
 }
 
@@ -124,14 +137,14 @@ body {
 .grid {
 	flex: 1;
 	display: grid;
-	grid-template-columns: repeat(auto-fill, minmax(432px, 1fr));
+	grid-template-columns: repeat(auto-fill, minmax(444px, 1fr));
 	grid-auto-rows: max-content;
 	grid-gap: 24px;
 	padding-top: 24px;
 }
 
 .templateItem {
-	width: 432px;
+	width: 444px;
 	height: 382px;
 	color: unset;
 	position: relative;
